EM.FDEM.SrcFDEM_XXX --> EM.FDEM.Src.XXX
