+++ SimPEG/EM/FDEM/__init__.py
@@ -1,3 +1,3 @@
-from SurveyFDEM import *
+from SurveyFDEM import RxFDEM, Src, SurveyFDEM
 from FDEM import BaseFDEMProblem, ProblemFDEM_e, ProblemFDEM_b, ProblemFDEM_j, ProblemFDEM_h
 from FieldsFDEM import *
